Deprecate Toshiba solvers
