--- azure-quantum/azure/quantum/optimization/__init__.py
@@ -1,11 +1,0 @@
-# coding=utf-8
-##
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-##
-
-from .term import *
-from .problem import *
-from .streaming_problem import *
-from .online_problem import *
-from azure.quantum.target.solvers import *
