--- azure-quantum/azure/quantum/optimization/toshiba/solvers.py
@@ -1,12 +1,0 @@
-# coding=utf-8
-##
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-##
-import warnings
-warnings.warn("The azure.quantum.optimization.toshiba.solvers namespace will be deprecated. \
-Please use azure.quantum.target.toshiba instead.")
-
-from azure.quantum.target.toshiba.solvers import (
-    SimulatedBifurcationMachine
-)
