+++ azure-quantum/tests.live/Run.ps1
@@ -55,9 +55,6 @@ function PyTestMarkExpr() {
     if ($AzureQuantumCapabilities -notcontains "submit.ionq") {
         $MarkExpr += " and not ionq"
     }
-    if ($AzureQuantumCapabilities -notcontains "submit.toshiba") {
-        $MarkExpr += " and not toshiba"
-    }
     if ($AzureQuantumCapabilities -notcontains "submit.rigetti") {
         $MarkExpr += " and not rigetti"
     }
