+++ azure-quantum/tests/unit/test_job_results.py
@@ -1,0 +1,73 @@
+##
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+##
+
+import unittest
+from unittest.mock import Mock
+
+from common import QuantumTestBase
+from azure.quantum import Job
+from azure.quantum._client.models import JobDetails
+
+
+class TestJobResults(QuantumTestBase):
+    """TestJob
+
+    Tests the azure.quantum.job module.
+    """
+
+    def test_job_success(self):
+        job_results = self._get_job_results("test_output_data_format","{\"Histogram\": [\"[0]\", 0.50, \"[1]\", 0.50]}")
+        self.assertTrue(len(job_results["Histogram"]) == 4)
+        
+
+    def test_job_for_microsoft_quantum_results_v1_success(self):
+        job_results = self._get_job_results("microsoft.quantum-results.v1","{\"Histogram\": [\"[0]\", 0.50, \"[1]\", 0.50]}")
+        self.assertTrue(len(job_results.keys()) == 2)
+        self.assertTrue(job_results["[0]"], 0.50)
+        self.assertTrue(job_results["[1]"], 0.50)
+
+    
+    def test_job_for_microsoft_quantum_results_v1_no_histogram_returns_raw_result(self):
+        job_result_raw = "{\"NotHistogramProperty\": [\"[0]\", 0.50, \"[1]\", 0.50]}"
+        job_result = self._get_job_results("microsoft.quantum-results.v1", job_result_raw)
+        self.assertTrue(job_result, job_result_raw)
+
+
+    def test_job_for_microsoft_quantum_results_v1_invalid_histogram_returns_raw_result(self):
+        job_result_raw = "{\"NotHistogramProperty\": [\"[0]\", 0.50, \"[1]\"]}"
+        job_result = self._get_job_results("microsoft.quantum-results.v1", job_result_raw)
+        self.assertTrue(job_result, job_result_raw)
+
+
+    def _get_job_results(self, output_data_format, results_as_json_str):
+        job_details = JobDetails(
+            id= "",
+            name= "",
+            provider_id="",
+            target="",
+            container_uri="",
+            input_data_format="",
+            output_data_format = output_data_format)
+        job_details.status = "Succeeded"
+        job = Job(
+            workspace=None, 
+            job_details=job_details)
+        
+        job.has_completed = Mock(return_value=True)
+        job.wait_until_completed = Mock()
+
+        class DowloadDataMock(object):
+            def decode(): str
+            pass
+
+        download_data = DowloadDataMock()
+        download_data.decode = Mock(return_value=results_as_json_str)
+        job.download_data = Mock(return_value=download_data)
+        
+        return job.get_results()
+
+
+if __name__ == "__main__":
+    unittest.main()
